Backup: Store the backup tree in a subvol
This makes it easier to delete.  We have a fancy "delete a whole
subvol recursively" script now.  The old setup of sometimes making the
root dir a subvol and sometimes not means cleanup is tricky.  In fact,
the cleanup script we're using now chokes it its told to cleanup a dir
that isn't a subvol.  Whoops.