Emacs: Replace org-sticky-header with my own version
The org-sticky-header package was nice, but I didn't like that it
didn't use org's own fontification logic.  That meant markup in header
titles would be displayed literally.  I also wanted to be able to
click on a heading title and jump back to that point.  Lastly, as a
bit of a stylistic choice, I want the header line aligned with the
margin rather than the buffer content start.  Since it isn't part of
the buffer, I think it looks nicer if it clearly breaks out of the box.